chore: move ticket index warning helpers to utils
