+++ packages/cli/src/utils/ticket-index-warnings.ts
